+++ Dev/DataModels/Atdi.DataModels.Sdrns.Device/Tasks/Signaling/ReferenceSituation.cs
@@ -9,6 +9,6 @@ namespace Atdi.DataModels.Sdrns.Device
     public class ReferenceSituation
     {
         public ReferenceSignal[] ReferenceSignal;
-        public int SensorId;
+        public long SensorId;
     }
 }
